chore: refactor perf-entry-recorder
* move bootstrap to the core
