+++ packages/rum/test/specs/apm-base.spec.js
@@ -25,8 +25,8 @@
 
 import ApmBase from '../../src/apm-base'
 import { createServiceFactory, PAGE_LOAD } from '@elastic/apm-rum-core'
+import { bootstrap } from '@elastic/apm-rum-core/src/bootstrap'
 import { TRANSACTION_END } from '@elastic/apm-rum-core/src/common/constants'
-import bootstrap from '../../src/bootstrap'
 import { getGlobalConfig } from '../../../../dev-utils/test-config'
 import Promise from 'promise-polyfill'
 import { scheduleTaskCycles } from '../../../rum-core/test'
